Implemented a two new AIs. Rewrote menu system. Added CLI.
Implemented a new AI which minimizes the number of tiles which
cannot be removed from the board. Used this new AI in combination
with our previous best-first search in order to create a third AI
which considers both the size of the gruop being removed and the
number of isolated tiles after the move. Completely rewrote the
menu system (so that the AIs could be moved to a submenu). Created
a command-line interface so that scripting could ease tournament
play. Added a tournament script which executes the combined AI for
each tournament board.