Condensed collapse into one function. Corrected seconds formatter.
I realized that I had previously included the collapse_rows
functionality into collapse_columns, so calling both was
literally wasting time. To correct this, I simply renamed
collapse_columns to collapse and removed the other methods. Also,
in quiet mode, the formatter for seconds was for a string type
rather than a float. The result was a loss of precision when the
float was printed. This was corrected.